OpenmrsGciBotApplication.java: programmatically set tomcat port
This is to fix the port binding error when app is packaged as container
and deployed to heroku. Heroku does not obey the docker EXPOSE command
in docker file and dynamically assigns a port number to deployed docker
webapps. Spring boot defaults to 8080 but heroku expects a different port
number and it's set in an environment variable $PORT. To fix this, we just
modify default tomcat port by setting it to the $PORT value

Closes https://github.com/ivange94/openmrs-gci-bot/issues/2